rename share to loading
