+++ ui/src/tugboat/components/index.ts
@@ -14,8 +14,7 @@ export * from './SelectedModelDescription';
 export * from './SelectExample';
 export * from './SelectModel';
 export * from './SelectModelAndDescription';
-export * from './share';
-export * from './shared';
+export * from './Loading';
 export * from './Spark';
 export * from './SyntaxHighlight';
 export * from './TaskDescription';
